+++ api/serializers.py
@@ -28,23 +28,6 @@ from threepio import logger
 
 
 # Serializers
-class MachineExportSerializer(serializers.ModelSerializer):
-    """
-    """
-    name = serializers.CharField(source='export_name')
-    instance = serializers.SlugRelatedField(slug_field='provider_alias')
-    status = serializers.CharField(default="pending")
-    disk_format = serializers.CharField(source='export_format')
-    owner = serializers.SlugRelatedField(slug_field='username',
-                                         source='export_owner')
-    file = serializers.CharField(read_only=True, default="",
-                                 required=False, source='export_file')
-
-    class Meta:
-        model = MachineExport
-        fields = ('id', 'instance', 'status', 'name',
-                  'owner', 'disk_format', 'file')
-
 class LicenseSerializer(serializers.ModelSerializer):
 
     created_by = serializers.SlugRelatedField(slug_field='username', read_only=True)
