+++ api/serializers/__init__.py
@@ -20,3 +20,4 @@ from .instance_serializer import InstanceSerializer
 from .instance_history_serializer import InstanceHistorySerializer
 from .paginated_instance_history_serializer import PaginatedInstanceHistorySerializer
 from .paginated_instance_serializer import PaginatedInstanceSerializer
+from .machine_export_serializer import MachineExportSerializer
