+++ api/serializers/machine_export_serializer.py
@@ -1,0 +1,20 @@
+from core.models.machine_export import MachineExport
+from rest_framework import serializers
+
+
+class MachineExportSerializer(serializers.ModelSerializer):
+    """
+    """
+    name = serializers.CharField(source='export_name')
+    instance = serializers.SlugRelatedField(slug_field='provider_alias')
+    status = serializers.CharField(default="pending")
+    disk_format = serializers.CharField(source='export_format')
+    owner = serializers.SlugRelatedField(slug_field='username',
+                                         source='export_owner')
+    file = serializers.CharField(read_only=True, default="",
+                                 required=False, source='export_file')
+
+    class Meta:
+        model = MachineExport
+        fields = ('id', 'instance', 'status', 'name',
+                  'owner', 'disk_format', 'file')
